% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_arguments.R
\name{compute_critical_t}
\alias{compute_critical_t}
\title{Title}
\usage{
compute_critical_t(df, sig_level, alternative = "two.sided")
}
\arguments{
\item{df}{numeric value}

\item{sig_level}{numeric value}

\item{alternative}{character value}
}
\value{
a numeric value
}
\description{
Title
}
