% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRDAbeta.R
\name{.onAttach}
\alias{.onAttach}
\title{Initial message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{used to define the libname}

\item{pkgname}{used to define the pkgname}
}
\value{
Initial message with suggestion to loook at vignettas
}
\description{
Initial message
}
