% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_tests.R
\name{my_cor_test}
\alias{my_cor_test}
\title{Title}
\usage{
my_cor_test(x, y, alternative = "two.sided", ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{alternative}{character value}

\item{...}{other options}
}
\value{
list with p.values and correlation estimate
}
\description{
Title
}
