% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_tests.R
\name{my_t_test}
\alias{my_t_test}
\title{my_t_test}
\usage{
my_t_test(
  x,
  y = NULL,
  test_method,
  alternative = "two.sided",
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{numeric value}

\item{y}{numeric value}

\item{test_method}{a character string}

\item{alternative}{character value}

\item{mu}{numeric value}

\item{paired}{logic value}

\item{var.equal}{logic value}

\item{conf.level}{numeric value}

\item{...}{other options}
}
\value{
list with p.values and Cohen's d estimate
}
\description{
my_t_test
}
