% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRDA-package.R
\docType{package}
\name{PRDA}
\alias{PRDA}
\title{PRDA: A Package for Prospective and Retrospective Design Analysis.}
\description{
The PRDA package provides two main important functions:
\code{prospective()} and \code{retrospective()}.
}
\section{PRDA functions}{

The PRDA functions ...
}

