% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospective.R
\name{prospective}
\alias{prospective}
\title{Title}
\usage{
prospective(
  effect_size,
  power,
  ratio_n2 = 1,
  effect_type = c("cohen_d", "correlation"),
  alternative = c("two.sided", "less", "greater"),
  sig_level = 0.05,
  B = 10000,
  seed = NULL,
  tl = -Inf,
  tu = Inf,
  B_effect = 250,
  sample_range = c(2, 1000),
  tol = 0.01,
  display_message = FALSE,
  ...
)
}
\arguments{
\item{effect_size}{numeric value or function}

\item{power}{numeric value}

\item{ratio_n2}{numeric value}

\item{effect_type}{character value.}

\item{alternative}{character value.}

\item{sig_level}{numeric value}

\item{B}{numeric value}

\item{seed}{numeric value}

\item{tl}{numeric value}

\item{tu}{numeric value}

\item{B_effect}{numeric value}

\item{sample_range}{two value numeric vector}

\item{tol}{numeric value}

\item{display_message}{logical}

\item{...}{optional argumets}
}
\value{
a list
}
\description{
Title
}
