% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective.R
\name{retrospective}
\alias{retrospective}
\title{Retrospective}
\usage{
retrospective(
  sample_group1,
  effect_size,
  sample_group2 = NULL,
  effect_type = c("cohen_d", "correlation"),
  alternative = c("two.sided", "less", "greater"),
  sig_level = 0.05,
  B = 1000,
  seed = NULL,
  ...
)
}
\arguments{
\item{sample_group1}{numeric value.}

\item{effect_size}{numeric value.}

\item{sample_group2}{numeric value.}

\item{effect_type}{character value.}

\item{alternative}{character value.}

\item{sig_level}{numeric value.}

\item{B}{numeric value.}

\item{seed}{numeric value.}

\item{...}{optional argument.}
}
\value{
a list
}
\description{
Retrospective
}
