% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_observations.R
\name{sample_obs_cor}
\alias{sample_obs_cor}
\title{Title}
\usage{
sample_obs_cor(sample_n1, effect_size)
}
\arguments{
\item{sample_n1}{numeric value}

\item{effect_size}{numeric value}
}
\value{
matrix of observations
}
\description{
Title
}
