% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{simple_sum}
\alias{simple_sum}
\title{An internal function}
\usage{
simple_sum(x, y)
}
\arguments{
\item{x}{numeric value}

\item{y}{numeric value}
}
\value{
the sum of the two values
}
\description{
An internal function
}
