% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_design_analysis.R
\name{summary.design_analysis}
\alias{summary.design_analysis}
\title{Print Method for design_analysis class}
\usage{
\method{summary}{design_analysis}(object, ...)
}
\arguments{
\item{object}{an object with class "design_analysis"}

\item{...}{urther arguments passed to or from other methods.}
}
\value{
a summary output
}
\description{
Print Method for design_analysis class
}
