% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_analysis.R
\name{analysis_cohen}
\alias{analysis_cohen}
\title{Title}
\usage{
analysis_cohen(sample_group1, sample_group2, effect_size, alternative, B, ...)
}
\arguments{
\item{sample_group1}{numeric value}

\item{sample_group2}{numeric value}

\item{effect_size}{numeric value}

\item{alternative}{character value}

\item{B}{numeric value}

\item{...}{other variables passed to}
}
\value{
a matrix
}
\description{
Title
}
