% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRDAbeta.R
\docType{package}
\name{PRDAbeta}
\alias{PRDAbeta}
\title{PRDAbeta: A package for cunduct prospective and retospective design analysis.}
\description{
The PRDAbeta package provides two main important functions:
\code{prospective()} and \code{retrospective()}.
}
\section{PRDAbeta functions}{

The PRDAbeta functions ...
}

