% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospective.R
\name{prospective}
\alias{prospective}
\title{Prospective Design Analysis}
\usage{
prospective(
  effect_size,
  power,
  ratio_n2 = 1,
  effect_type = c("cohen_d", "correlation"),
  alternative = c("two.sided", "less", "greater"),
  sig_level = 0.05,
  B = 10000,
  seed = NULL,
  tl = -Inf,
  tu = Inf,
  B_effect = 1000,
  sample_range = c(2, 1000),
  tol = 0.01,
  display_message = TRUE,
  ...
)
}
\arguments{
\item{effect_size}{a numeric value or function (see details) indicating the
hypothetical population effect size.}

\item{power}{a numeric value indicating the required power level.}

\item{ratio_n2}{a numeric value indicating the ratio between \code{sample_n1}
and \code{sample_n2}.}

\item{effect_type}{a character string specifying the effect type, must be one
of "cohen_d" (default, Cohen's \emph{d} standardised means difference) or
"pearson" (Pearson's correlation). You can specify just the initial letter.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify
just the initial letter.}

\item{sig_level}{a numeric value indicating the significance level on which
the alternative hypothesis is evaluated.}

\item{B}{a numeric  value indicating the number of iterations. Increase the
number of iterations to obtain more stable results.}

\item{seed}{a numeric value indicating the seed for random number generation.
Set the seed to obtain reproducible results.}

\item{tl}{optional value indicating the lower truncation point if
\code{effect_size} is defined as a function.}

\item{tu}{optional value indicating the upper truncation point if
\code{effect_size} is defined as a function.}

\item{B_effect}{a numeric  value indicating the number of sampled effects
if \code{effect_size} is defined as a function. Increase the number to
obtain more stable results.}

\item{sample_range}{a length-2 numeric vector indicating the minimum and
maximum sample size of the first group (\code{sample_n1}).}

\item{tol}{a numeric value indicating the tolerance of required power level.}

\item{display_message}{a logical variable indicating whether to display or
not the information about computational steps.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
A list with class "design_analysis" containing the following
 components:
   \item{design_analysis}{a character string indicating the type of design
   analysis: "prospective".}
   \item{call_arguments}{a list with all the arguments passed to the
   function.}
   \item{effect_info}{a list with all the information regarding the
   considered hypothetical population effect size. The list includes:
   \code{effect_type} indicating the type of effect; \code{effect_function}
   indicating the function from which effect are sampled or the string
   "single_value" if a single value was provided; \code{effect_summary}
   summary of the sampled effects; \code{effect_samples} vector with the
   sampled effects (or unique value in the case of a single value); if
   relevant \code{tl} and \code{tu} specifying the lower upper truncation
   point respectively.}
   \item{test_info}{a list with all the information regarding the test
   performed. The list includes: \code{test_method} character sting
   indicating the test method (e.g., "pearson", "one-sample", "paired",
   "two-samples", or "welch"); the required sample size (\code{sample_n1} and
   if relevant \code{sample_n2}), the alternative hypothesis
   (\code{alternative}), significance level (\code{sig_level})  and  degrees
   of freedom (\code{df}) of the statistical test; \code{critical_effect} the
   minimum absolute effect value that would result significant. Note that
   \code{critical_effect} in the case of \code{alternative = "two.sided"} is
   the absolute value and both positive and negative values should be
   considered.}
   \item{prospective_res}{a data frame with the results of the design
   analysis. Columns names are \code{power}, \code{typeM}, and \code{typeS}.}
}
\description{
Given the hypothetical population effect size and the required power level,
the function \code{prospective()} performs a prospective design analysis for
Cohen's \emph{d} (\emph{t}-test comparing group means) or  Pearson's
correlation test between two variables. According to the defined alternative
hypothesis and the significance level, the required sample size is computed
together with the associated Type M error, Type S error, and the critical
correlation value (i.e., the minimum absolute effect size value that would
result significant).
}
\details{
Conduct a prospective design analysis to define the required sample
  size and the associated inferential risks according to study design. A
  general overview is provided in the \code{vignette("prospective")}.

  \strong{Population effect size}

  The hypothetical population effect size (\code{effect_size}) can be set to
  a single value or a function that allows to sample values from a given
  distribution. The function has to be defined as \code{function(x)
  my_function(x, ...)}, with only one single variable \code{x} that represent
  the number of samples (e.g., \code{function(x) rnorm(x, mean = 0, sd = 1)};
  \code{function(x) sample(c(.1,.3,.5), x, replace = TRUE)}). This allows
  users to define hypothetical effect size distribution according to their
  needs.

  Argument \code{B_effect} allows defining the number of sampled effects.
  Users can access sampled effects in the \code{effect_info} list included in
  the output to evaluate if the sample is representative of their
  specification. Increase the number to obtain more accurate results but it
  will require more computational time (default is 1000). To avoid long
  computational times, we suggest adjusting \code{B} when using a function to
  define the hypothetical population effect size.

  Optional arguments \code{tl} and \code{tu} allow truncating the sampling
  distribution specifying the lower truncation point and upper truncation
  point respectively. Note that if \code{effect_type = "correlation"},
  distribution is automatically truncated between -1 and 1.

  \strong{Effect type options}

  The \code{effect_type} argument can be set to \code{"cohen_d"} (default)
  for standardized mean difference or \code{"correlation"} if Pearson's
  correlation is evaluated.

  In the case of \code{"cohen_d"} one-sample or two-samples \emph{t}-test are
  considered following same options specification of basic function
  \code{t.test()}, note that default options of \code{t.test()} are
  \code{paired = FALSE} and \code{var.equal = FALSE}. For one-sample
  \emph{t}-test \code{ratio_n = NULL} is required. For paired \emph{t}-test
  \code{ratio_n = 1} and option \code{paired = TRUE} are required. For
  two-samples \emph{t}-test \code{ratio_n} can be specified according to user
  needs and option \code{var.equal = TRUE} is required. For Welch
  \emph{t}-test, \code{ratio_n} can be specified according to user needs
  (default option is \code{var.equal = FALSE}).

  In the case of \code{"correlation"}, only Pearson's correlation between two
  variables is available and \code{ratio_n} is set to 1 (default). The
  Kendall's \emph{tau} and Spearman's \emph{rho} are not implemented.

  \strong{Study design}

  Study design can be further defined according to statistical test
  directionality and required \eqn{\alpha}-level using the arguments
  \code{alternative} and \code{sig_level} respectively.
}
\examples{

# One-sample t-test
prospective(effect_size = .3, power = .8, ratio_n = NULL,
            effect_type = "cohen_d", seed = 2020, B = 1e3)
# Paired t-test
prospective(effect_size = .3, power = .8, ratio_n = 1,
            effect_type = "cohen_d", paired = TRUE, seed = 2020, B = 1e3)
# Two-samples t-test
prospective(effect_size = .3, power = .8, ratio_n = 1.5,
            effect_type ="cohen_d", var.equal = TRUE, seed = 2020, B = 1e3)
# Welch t-test
prospective(effect_size = .3, power = .8, ratio_n = 2,
            effect_type ="cohen_d", seed = 2020, B = 1e3)

# Pearson's correlation
prospective(effect_size = .3, power = .8, effect_type = "correlation",
            seed = 2020, B = 1e3)

\dontrun{
# Define effect_size using functions (long computational time)
prospective(effect_size = function(x) rnorm(x, .3, .1), power = .8,
              effect_type = "correlation", seed = 2020)
prospective(effect_size = function(x) rnorm(x, .3, .1), power = .8,
              effect_type = "cohen_d", tl = .2, tu = .4, seed = 2020)
}

}
\references{
Altoè, G., Bertoldo, G., Zandonella Callegher, C., Toffalini, E.,
 Calcagnì, A., Finos, L., & Pastore, M. (2020). Enhancing Statistical
 Inference in Psychological Research via Prospective and Retrospective Design
 Analysis. Frontiers in Psychology, 10.
 \url{https://doi.org/10.3389/fpsyg.2019.02893}

 Gelman, A., & Carlin, J. (2014). Beyond Power Calculations: Assessing Type S
 (Sign) and Type M (Magnitude) Errors. Perspectives on Psychological Science,
 9(6), 641–651. \url{https://doi.org/10.1177/1745691614551642}

 Bertoldo, G., Altoè, G., & Zandonella Callegher, C. (2020, June 15).
 Designing Studies and Evaluating Research Results: Type M and Type S Errors
 for Pearson Correlation Coefficient. Retrieved from \url{https://psyarxiv.com/q9f86/}
}
