% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective.R
\name{retrospective}
\alias{retrospective}
\title{Retrospective}
\usage{
retrospective(
  sample_n1,
  effect_size,
  sample_n2 = NULL,
  effect_type = c("cohen_d", "correlation"),
  alternative = c("two.sided", "less", "greater"),
  sig_level = 0.05,
  B = 10000,
  seed = NULL,
  tl = -Inf,
  tu = Inf,
  B_effect = 250,
  ...
)
}
\arguments{
\item{sample_n1}{numeric value.}

\item{effect_size}{numeric value or function.}

\item{sample_n2}{numeric value.}

\item{effect_type}{character value.}

\item{alternative}{character value.}

\item{sig_level}{numeric value.}

\item{B}{numeric value.}

\item{seed}{numeric value.}

\item{tl}{numeric value.}

\item{tu}{numeric value.}

\item{B_effect}{numeric value.}

\item{...}{optional argument.}
}
\value{
a list
}
\description{
Retrospective
}
