% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_analysis.R
\name{retrospective_cohen}
\alias{retrospective_cohen}
\title{Title}
\usage{
retrospective_cohen(
  sample_n1,
  sample_n2,
  effect_size,
  test_method,
  alternative,
  sig_level,
  B,
  ...
)
}
\arguments{
\item{sample_n1}{numeric value}

\item{sample_n2}{numeric value}

\item{effect_size}{numeric value}

\item{test_method}{character value}

\item{alternative}{character value}

\item{sig_level}{numeric value.}

\item{B}{numeric value}

\item{...}{other variables passed to}
}
\value{
a matrix
}
\description{
Title
}
